jsii-python/requirements.txt
aws-jsiisamples.jsii-code-samples==1.7.0
